start refactoring code
