+++ pages/[date]/[room]/index.js
@@ -1,0 +1,9 @@
+function Room() {
+  return (
+    <>
+      <h1>Hello</h1>
+    </>
+  );
+}
+
+export default Room;
